+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ _Date TBD_
 
 ### New features
 * Added option to download csv summaries of 2014/2015 products to resources page
+* Added Developer status display / edit
 
 ---